pyproject.toml
[tool.poetry]
name = "drf-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.8"
django-filter = "^24.3"
pillow = "^11.0.0"
django-money = "^3.5.3"
django-phonenumber-field = "^8.0.0"
cors = "^1.0.1"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
